Reject non-canonical field elements at deserialization boundaries
Close the snarkjs CVE-2023-33252 acceptance class (roadmap 1.7): a field
element at or above its modulus is a second byte encoding of the same
residue and now rejects at the boundary instead of flowing into
transcript absorbs and reduced arithmetic.

New public WellKnownCurves.IsCanonicalScalar (allocation-free big-endian
compare against pre-calculated per-curve order bytes) backs checks at:

- R1csMatrix.FromSortedTriples and RawR1csWitness.FromCanonical, the two
  funnel factories every Circom/ZkInterface reader path flows through.
- RawR1csInstance.Create public inputs; RelaxedR1csInstance.Create
  public inputs and u (both transcript-absorbed).
- BbsSignature/BbsProof.FromCanonical: every scalar slot in [1, r-1] per
  the spec octets_to_signature/octets_to_proof.
- SecdsaSignature.FromCanonical r, s in [1, n-1] (the documented
  invariant); DlEqualityProof.FromCanonical s only (r is the full-width
  Fiat-Shamir digest by design).
- LongfellowCircuitReader.TryRead gains an opt-in constant-table range
  delegate; the mdoc facade wires InRangeFp256 for the signature circuit.

Also fix the unchecked-uint wrap in the Circom .r1cs header consistency
check (nPubOut + nPubIn + nPrvIn + 1 computed in ulong).

Table-driven rejection tests pin every boundary (order-1 accepts, order
and above rejects, messages pinned), including a facade-level pin that
the Longfellow guard stays wired. The masked-Spartan soundness sampler
now draws canonical fake witnesses via DeterministicScalarFill.

Deliberately deferred to W1-b: scalar-section canonicity for the Spartan
proof-container family (SpartanProof et al., RangeProof, BaseFold
serialization) - whether Fiat-Shamir absorbs every scalar section is
exactly what the transcript-completeness audit decides, and the section
maps get built there. BBS wrong-subgroup point rejection (KEY-VALIDATE /
octets_to_point subgroup checks) is a public-API-shape decision reported
separately.

+++ src/Lumoin.Veridical.Core/ConstraintSystems/Interop/Circom/CircomR1csReader.cs
@@ -286,7 +286,10 @@ public static class CircomR1csReader
             throw new ArgumentException("R1CS header declares nConstraints = 0; constraint section is mandatory.");
         }
 
-        if(nPubOut + nPubIn + nPrvIn + 1 > nWires)
+        //The sum is computed in ulong: in uint arithmetic a crafted header
+        //(e.g. nPubOut = 0xFFFFFFFF, nPubIn = 1) wraps mod 2^32 and slips
+        //past this consistency check.
+        if((ulong)nPubOut + nPubIn + nPrvIn + 1 > nWires)
         {
             throw new ArgumentException(
                 $"R1CS header inconsistency: nPubOut ({nPubOut}) + nPubIn ({nPubIn}) + nPrvIn ({nPrvIn}) + 1 (constant) exceeds nWires ({nWires}).");
